refactor: component を1箇所にまとめるため
